requirements.txt
pymongo>=4.6.0
requests>=2.32.0
tqdm>=4.66.0
pandas>=2.2.0
gitpython>=3.1.40
python-dotenv>=1.0.1
